requirements.txt
asgiref==3.6.0
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.1.6
django-extensions==3.2.1
django-filter==23.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django_debug_toolbar==3.8.1
executing==1.2.0
ipython==8.9.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
tzdata==2022.7
wcwidth==0.2.6
